+++ Modele/Agents/Retina.cpp
@@ -2,7 +2,7 @@
 
 using namespace sim;
 
-Retina::Retina(uint32_t seg, float depth, float fov)
+Retina::Retina(uint32_t seg, double depth, double fov)
     : _nb_segments(seg),
       _field_of_view(fov),
       _depth(depth),
